pyproject.toml
[project]
name = "stream-summarization"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "openai>=1.93.0",
    "psycopg2-binary>=2.9.10",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "pydantic-settings>=2.10.1",
    "langchain-openai>=0.3.7",
    "fpdf2>=2.8.3",
    "python-docx>=1.1.2",
    "pypdf>=5.1.0",
    "odfpy>=1.4.1",
    "requests>=2.32.4",
    "langchain>=0.3.27",
]

[dependency-groups]
dev = [
    "printdirtree>=0.1.5",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
